Hoist type converter into separate class/file `QuintTypeConverter`
